refactor(l1): rename ExecutionDB to ProverDB.
