+++ crates/l2/prover/zkvm/interface/src/lib.rs
@@ -24,7 +24,7 @@ pub mod io {
         types::{Block, BlockHeader},
         H256,
     };
-    use ethrex_vm::ExecutionDB;
+    use ethrex_vm::ProverDB;
     use serde::{de::DeserializeOwned, Deserialize, Serialize};
     use serde_with::{serde_as, DeserializeAs, SerializeAs};
 
@@ -39,7 +39,7 @@ pub mod io {
         #[serde_as(as = "SerdeJSON")]
         pub parent_block_header: BlockHeader,
         /// database containing only the data necessary to execute
-        pub db: ExecutionDB,
+        pub db: ProverDB,
         /// value used to calculate base fee
         pub elasticity_multiplier: u64,
     }
@@ -105,7 +105,7 @@ pub mod trie {
     use ethrex_rlp::{decode::RLPDecode, encode::RLPEncode, error::RLPDecodeError};
     use ethrex_storage::{hash_address, hash_key, AccountUpdate};
     use ethrex_trie::{Trie, TrieError};
-    use ethrex_vm::ExecutionDB;
+    use ethrex_vm::ProverDB;
     use thiserror::Error;
 
     #[derive(Debug, Error)]
@@ -121,7 +121,7 @@ pub mod trie {
     }
 
     pub fn verify_db(
-        db: &ExecutionDB,
+        db: &ProverDB,
         state_trie: &Trie,
         storage_tries: &HashMap<H160, Trie>,
     ) -> Result<bool, Error> {
