+++ crates/vm/backends/levm/db.rs
@@ -4,7 +4,8 @@ use ethrex_common::{Address as CoreAddress, H256 as CoreH256};
 use ethrex_levm::constants::EMPTY_CODE_HASH;
 use ethrex_levm::db::Database as LevmDatabase;
 
-use crate::db::{ExecutionDB, StoreWrapper};
+use crate::db::StoreWrapper;
+use crate::ProverDB;
 use ethrex_levm::db::error::DatabaseError;
 use std::collections::HashMap;
 use std::result::Result;
@@ -160,7 +161,7 @@ impl LevmDatabase for StoreWrapper {
     }
 }
 
-impl LevmDatabase for ExecutionDB {
+impl LevmDatabase for ProverDB {
     fn get_account(&self, address: CoreAddress) -> Result<Account, DatabaseError> {
         let Some(acc_info) = self.accounts.get(&address) else {
             return Ok(Account::default());
