+++ crates/vm/levm/bench/revm_comparison/src/lib.rs
@@ -9,7 +9,7 @@ use ethrex_levm::{
     vm::VM,
     Environment,
 };
-use ethrex_vm::db::ExecutionDB;
+use ethrex_vm::ProverDB;
 use revm::{
     db::BenchmarkDB,
     primitives::{address, Address, Bytecode, TransactTo},
@@ -55,13 +55,13 @@ pub fn run_with_levm(program: &str, runs: u64, calldata: &str) {
         ),
     ];
 
-    let mut execution_db = ExecutionDB::default();
+    let mut prover_db = ProverDB::default();
 
     accounts.iter().for_each(|(address, account)| {
-        execution_db.accounts.insert(*address, account.info.clone());
+        prover_db.accounts.insert(*address, account.info.clone());
     });
 
-    let mut db = GeneralizedDatabase::new(Arc::new(execution_db), CacheDB::new());
+    let mut db = GeneralizedDatabase::new(Arc::new(prover_db), CacheDB::new());
 
     cache::insert_account(
         &mut db.cache,
