+++ crates/vm/lib.rs
@@ -3,11 +3,13 @@ pub mod db;
 mod errors;
 mod execution_result;
 mod helpers;
+mod prover_db;
 
 pub mod backends;
 
 pub use backends::{BlockExecutionResult, Evm, EvmEngine};
-pub use db::{ExecutionDB, StoreWrapper};
-pub use errors::{EvmError, ExecutionDBError};
+pub use db::StoreWrapper;
+pub use errors::{EvmError, ProverDBError};
 pub use execution_result::ExecutionResult;
 pub use helpers::{create_contract_address, fork_to_spec_id, SpecId};
+pub use prover_db::ProverDB;
